rules: Add RHEL 7.0 to UseSystem{Boost,TBB}
In rules.py, there are two functions UseSystemBoost and UseSystemTBB.
They evaluate the build hosts distribution and release level against a
hardcoded list, and if there is a match (i.e. the build host's
distribution matches an entry in the list AND the release level is equal
to or greater than the matching entry), the function will return True.
Otherwise, it returns false.

The implication is that if a distribution is listed, system
develop-libraries will be used for both boost and tbb.

Without this change, builds have begun failing on up-to-date RHEL 7.2
machines.

Change-Id: I85a88337e091e03575ab44c6ae4746444aeff834
Closes-Bug: #1554553
Closes-Bug: #1554648
Signed-off-by: Martin Millnert <martin@millnert.se>