Break the buildinfo.h->buildinfo.h dependency.
We are creating circular dependencies (direct and indirect) by passing
source to the env.Command that's used to generate build information.

This is not passed as a functional dependency (as-in, source information
is not used for the build info itself) and it's used as a way to force
SCons to regenerate those files on every build.

Pass an empty source list to env.Command, and wrap output in
env.AlwaysRebuild.

Change-Id: I009204e9bc6f80197f16bde248bbef254d12ab9c