Fix setup_venv code to unblock unit tests
setup_venv is a very quirky function that is downloading random tarballs
from the internet for the sake of running some python tests. This has to
be rewritten to not depend on stuff that is not installed in the system,
but for now put a bandaid on it to unblock tests.

Change-Id: I78e159cc501739bfc034efd8115e9c10fc8f5e41
(cherry picked from commit b1f341466213049be68a344581a341dce1f0984b)