Enhance GoSetupCommon optionally change workdir
and retry specified number of times.

Made GoBuild, GoTest as scons Builders, So that the they
will not be built when other scons target is built.
For example currently GoBuild/GoTest are functions and so
they are invoked while building other scons targets such
as controller/src/config/*. To aovid this they are changed
to scons Builders

Enhanced GoTest to find the list of go test files

Added GoMod scons Builder to cache the go dependencies
using go mod tidy or go mod download commands.

Closes-Jira-Bug: CEM-13091
Change-Id: Iaa7f5df1749e726ef118296c25ddedf99ccb275c
Depends-On: Ib5139a32b86062fe2e9c2cf28e578ba408e9118d